Handled invalid credentials error scenario
In case of invalid credentials, a server response is not recieved rather, the JWT library throws an error. Hence, error handling has been performed to cater to this case. This further necessitated altering the failed login attempt unit test - simulation of an observable error that is thrown rather returning of some valid response.